+++ packages/editor/web/src/RichContentEditor/RichContentEditor.tsx
@@ -206,6 +206,8 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
 
   editorWrapper!: Element;
 
+  TiptapEditor: React.FC<{ editorProps; onUpdate }> | null;
+
   lastFocusedAtomicPlugin?: ContentBlock;
 
   updateBounds!: (editorBounds?: BoundingRect) => void;
@@ -264,7 +266,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
       undoRedoStackChanged: false,
     };
     this.refId = Math.floor(Math.random() * 9999);
-
+    this.TiptapEditor = null;
     this.commonPubsub = simplePubsub();
     this.handleCallbacks = this.createContentMutationEvents(
       this.state.editorState,
@@ -292,12 +294,26 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     this.handleBlockFocus(this.state.editorState);
   }
 
+  tiptapInterval;
+
   componentDidMount() {
     preventWixFocusRingAccessibility(this.editorWrapper);
     this.reportDebuggingInfo();
     this.preloadLibs();
     document?.addEventListener('beforeinput', this.preventDefaultKeyCommands);
     this.commonPubsub.set('undoExperiment', this.getUndoExperiment);
+
+    if (this.props.experiments?.tiptapEditor?.enabled) {
+      this.tiptapInterval = setTimeout(async () => {
+        const tiptapEditorModule = await import(
+          /* webpackChunkName: wix-tiptap-editor */
+          'wix-tiptap-editor'
+        );
+        const { TiptapEditor } = tiptapEditorModule;
+        this.TiptapEditor = TiptapEditor;
+        this.forceUpdate();
+      }, 1000);
+    }
   }
 
   componentWillMount() {
@@ -307,6 +323,7 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
   }
 
   componentWillUnmount() {
+    clearTimeout(this.tiptapInterval);
     this.updateBounds = () => '';
     document?.removeEventListener('beforeinput', this.preventDefaultKeyCommands);
   }
@@ -963,6 +980,15 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     return handled || 'not-handled';
   };
 
+  renderTiptapEditor = () => {
+    if (this.TiptapEditor) {
+      const TiptapEditor = this.TiptapEditor;
+      return <TiptapEditor onUpdate={() => null} editorProps={this.props} />;
+    } else {
+      return null;
+    }
+  };
+
   renderEditor = () => {
     const {
       editorKey,
@@ -1175,6 +1201,8 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
         [styles.desktop]: !isMobile,
         ...themeDesktopStyle,
       });
+
+      const isTiptapEditor = this.props.experiments?.tiptapEditor?.enabled;
       return (
         <GlobalContext.Provider value={this.state.context}>
           <Measure bounds onResize={this.onResize}>
@@ -1196,7 +1224,13 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
                   style={editorStyle}
                 >
                   {this.renderAccessibilityListener()}
-                  {this.renderEditor()}
+
+                  {isTiptapEditor && this.renderTiptapEditor()}
+                  {isTiptapEditor && (
+                    <div style={{ position: 'absolute', zIndex: -1 }}> {this.renderEditor()}</div>
+                  )}
+
+                  {!isTiptapEditor && this.renderEditor()}
                   {showToolbars && this.renderToolbars()}
                   {this.renderInlineModals()}
                   {this.renderErrorToast()}