+++ packages/ui-components/web/src/consts.ts
@@ -35,9 +35,9 @@ export const FOOTER_BUTTON_ALIGNMENT = Object.freeze({
   END: 'end',
 });
 
-// export const MODAL_CONTROLS_POSITION = Object.freeze({
-//   TOP: 'top',
-//   BOTTOM: 'bottom',
-// });
+export const MODAL_CONTROLS_POSITION = Object.freeze({
+  TOP: 'top',
+  BOTTOM: 'bottom',
+});
 
 // export const UNSUPPORTED_BLOCKS_TYPE = 'unsupported-blocks';