+++ CHANGELOG.md
@@ -17,6 +17,13 @@
     Click to see more.
   </summary>
 
+### :house: Internal
+- `plugin-commons`
+  - [#2608](https://github.com/wix-incubator/rich-content/pull/2608) moved SettingsMobileHeader/UrlInputModal components into ui-components folder
+- `tiptap-editor`
+  - [#2546](https://github.com/wix-incubator/rich-content/pull/2546) package added
+  
+
 </details>
 <hr/>
 
@@ -27,6 +34,8 @@
 - `editor`
   - [#2593](https://github.com/wix/ricos/pull/2593) fix cut behavior
   - [#2603](https://github.com/wix/ricos/pull/2603) fix insertDecoration (link) from Editor Commands
+- `link`
+  - [#2586](https://github.com/wix/ricos/pull/2586) support anchor from outer editor into collapsible list
 - `button`
   - [#2609](https://github.com/wix/ricos/pull/2609) Default Colors were always shown on settings panel configurations